Refactored to use common mapTopProps
