+++ src/routes/details/DetailsTableExpand.tsx
@@ -1,34 +1,19 @@
 import { Bullseye, Spinner } from '@patternfly/react-core';
 import { Td, Th, Tr } from '@patternfly/react-table';
 import type { Query } from 'api/queries';
-import { getQuery, parseQuery } from 'api/queries';
-import type { Report } from 'api/reports/report';
-import { ReportPathsType, ReportType } from 'api/reports/report';
-import type { AxiosError } from 'axios';
 import messages from 'locales/messages';
-import { cloneDeep } from 'lodash';
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import type { WrappedComponentProps } from 'react-intl';
 import { injectIntl } from 'react-intl';
-import { useDispatch, useSelector } from 'react-redux';
 import type { RouteComponentProps } from 'react-router-dom';
 import { withRouter } from 'react-router-dom';
-import type { AnyAction } from 'redux';
-import type { ThunkDispatch } from 'redux-thunk';
-import { getDateRange, getDateRangeDefault } from 'routes/utils/dateRange';
-import type { RootState } from 'store';
 import { FetchStatus } from 'store/common';
-import { reportActions, reportSelectors } from 'store/reports';
 import { getUnsortedComputedReportItems } from 'utils/computedReport/getComputedReportItems';
 import { formatCurrency } from 'utils/format';
 
-import { accountData } from './data/accountData';
-import { affiliateData } from './data/affiliateData';
-import { productData } from './data/productData';
-import { sourceData } from './data/sourceData';
+import { detailsMapToProps } from './api';
 import type { DetailsTableCell } from './DetailsTable';
 import { reportItem, reportItemValue } from './DetailsTable';
-import { GroupByType } from './utils';
 
 interface DetailsTableExpandOwnProps {
   columns?: any[];
@@ -36,20 +21,7 @@ interface DetailsTableExpandOwnProps {
   groupBy?: string;
   groupByValue?: string;
   isExpanded?: boolean;
-  isHidden?: boolean;
-  rowIndex?: number;
   secondaryGroupBy?: string;
-  setRowIndex?: (rowIndex: number, callback) => any;
-}
-
-interface DetailsTableExpandStateProps {
-  end_date?: string;
-  query: Query;
-  queryString: string;
-  report: Report;
-  reportError: AxiosError;
-  reportFetchStatus: FetchStatus;
-  start_date?: string;
 }
 
 type DetailsTableExpandProps = DetailsTableExpandOwnProps & RouteComponentProps<void> & WrappedComponentProps;
@@ -68,9 +40,6 @@ export const baseQuery: Query = {
   },
 };
 
-const reportPathsType = ReportPathsType.details; // Todo: temporary placeholder for upcoming API
-const reportType = ReportType.cost;
-
 const DetailsTableExpandBase: React.FC<DetailsTableExpandProps> = ({
   columns,
   dateRange,
@@ -81,7 +50,21 @@ const DetailsTableExpandBase: React.FC<DetailsTableExpandProps> = ({
   secondaryGroupBy,
 }) => {
   const [rows, setRows] = useState([]);
-  const { report, reportFetchStatus } = mapToProps({ dateRange, groupBy, groupByValue, secondaryGroupBy });
+  const { report, reportFetchStatus } = detailsMapToProps({
+    dateRange,
+    groupBy,
+    groupByValue,
+    isExpanded,
+    secondaryGroupBy,
+  });
+
+  const isMounted = useRef(false);
+  useEffect(() => {
+    isMounted.current = true;
+    return () => {
+      isMounted.current = false;
+    };
+  }, []);
 
   const initDatum = () => {
     if (!report) {
@@ -133,8 +116,10 @@ const DetailsTableExpandBase: React.FC<DetailsTableExpandProps> = ({
   };
 
   useEffect(() => {
-    initDatum();
-  }, [report]);
+    if (isMounted.current) {
+      initDatum();
+    }
+  }, [dateRange, secondaryGroupBy]);
 
   return (
     <React.Fragment>
@@ -150,115 +135,28 @@ const DetailsTableExpandBase: React.FC<DetailsTableExpandProps> = ({
         </Tr>
       ) : (
         rows.map((cells, rowIndex) => (
-          <React.Fragment key={`row-${rowIndex}`}>
-            <Tr isExpanded={isExpanded}>
-              {cells.map((item, cellIndex) =>
-                cellIndex === 0 ? (
-                  <Th
-                    dataLabel={columns[cellIndex]}
-                    key={`expanded-cell-${cellIndex}-${rowIndex}`}
-                    hasRightBorder
-                    isStickyColumn
-                  >
-                    <span className="expandedCol">{item.value}</span>
-                  </Th>
-                ) : (
-                  <Td dataLabel={columns[cellIndex]} key={`cell-${rowIndex}-${cellIndex}`}>
-                    {item.value}
-                  </Td>
-                )
-              )}
-            </Tr>
-          </React.Fragment>
+          <Tr key={`row-${rowIndex}`} isExpanded={isExpanded}>
+            {cells.map((item, cellIndex) =>
+              cellIndex === 0 ? (
+                <Th
+                  dataLabel={columns[cellIndex]}
+                  key={`expanded-cell-${cellIndex}-${rowIndex}`}
+                  hasRightBorder
+                  isStickyColumn
+                >
+                  <span className="expandedCol">{item.value}</span>
+                </Th>
+              ) : (
+                <Td dataLabel={columns[cellIndex]} key={`cell-${rowIndex}-${cellIndex}`}>
+                  {item.value}
+                </Td>
+              )
+            )}
+          </Tr>
         ))
       )}
     </React.Fragment>
   );
-};
-
-const mapToProps = ({
-  dateRange,
-  groupBy,
-  groupByValue,
-  secondaryGroupBy,
-}: DetailsTableExpandOwnProps): DetailsTableExpandStateProps => {
-  const dispatch: ThunkDispatch<RootState, any, AnyAction> = useDispatch();
-
-  const queryFromRoute = parseQuery<Query>(location.search);
-  const _dateRange = dateRange || getDateRangeDefault(queryFromRoute);
-  const { end_date, start_date } = getDateRange(_dateRange);
-
-  const query = {
-    filter: {
-      [groupBy]: groupByValue,
-    },
-    filter_by: queryFromRoute.filter_by || baseQuery.filter_by,
-    group_by: secondaryGroupBy,
-    order_by: queryFromRoute.order_by,
-  };
-  const queryString = getQuery({
-    ...query,
-    dateRange: undefined,
-    end_date,
-    start_date,
-  });
-
-  const report = useSelector((/* state: RootState */) => {
-    // reportSelectors.selectReport(state, widget.reportPathsType, widget.reportType, queryString)
-
-    let result;
-    switch (secondaryGroupBy) {
-      case GroupByType.affiliate:
-        result = cloneDeep(affiliateData);
-        break;
-      case GroupByType.account:
-        result = cloneDeep(accountData);
-        break;
-      case GroupByType.sourceOfSpend:
-        result = cloneDeep(sourceData);
-        break;
-      case GroupByType.product:
-      default:
-        result = cloneDeep(productData);
-        break;
-    }
-
-    const startDate = new Date(start_date + 'T23:59:59z');
-    const endDate = new Date(end_date + 'T23:59:59z');
-    endDate.setMonth(endDate.getMonth() + 1);
-
-    result.data = result.data.filter(item => {
-      const currentDate = new Date(item.date + 'T23:59:59z');
-      if (currentDate >= startDate && currentDate <= endDate) {
-        return item;
-      }
-    });
-
-    return result;
-  });
-  const reportFetchStatus = useSelector((state: RootState) =>
-    reportSelectors.selectReportFetchStatus(state, reportPathsType, reportType, queryString)
-  );
-  const reportError = useSelector((state: RootState) =>
-    reportSelectors.selectReportError(state, reportPathsType, reportType, queryString)
-  );
-
-  useMemo(() => {
-    if (reportFetchStatus !== FetchStatus.inProgress) {
-      // Todo: simulate time to fetch
-      dispatch(reportActions.fetchReport(reportPathsType, reportType, queryString));
-    }
-  }, [queryString]);
-
-  return {
-    end_date,
-    query,
-    queryString,
-    report,
-    reportFetchStatus,
-    reportError,
-    start_date,
-  };
 };
 
 const DetailsTableExpand = injectIntl(withRouter(DetailsTableExpandBase));
